# APP03022-mrch-bff-somx-ppsomx-util/cloud-endpoint/openapi.yaml
swagger: '2.0'
info:
  contact:
    email: jrcastillo@falabella.cl
    name: Arquitectura Falabella Tecnologia Corp.
  description: |
    Support RESTful operations for utils flow. This API provides access to parameters, modules, messages, application messages, processes, item types, items, shared catalogs, suppliers, supplier blocks, status train, conversions and catalog layout validation.
  title: Utils API
  version: 2.0.0
  x-business-acronym: mrch
  x-capability-acronym: utils
  x-consumer-segments: retail
  x-service-name: utils
host: '${DOMAIN_OPENAPI}'
schemes:
  - https
paths:
...
  /security/profile-users:
    get:
      tags:
        - security
      summary: Search profile-user summary
      operationId: searchProfileUsers
      produces:
        - application/json
      parameters:
        - name: startDate
          in: query
          required: true
          type: string
          format: date
        - name: endDate
          in: query
          required: true
          type: string
          format: date
        - name: entityId
          in: query
          required: false
          type: string
        - name: entityName
          in: query
          required: false
          type: string
        - name: status
          in: query
          required: false
          type: integer
          format: int32
        - name: langId
          in: query
          required: false
          type: integer
          format: int32
      responses:
        '200':
          description: Profile-user summary list
          schema:
            $ref: '#/definitions/SecuritySummaryListResponse'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ExceptionWrapper'
      security:
        - bearerAuth: []
  /security/profile-modules:
    get:
      tags:
        - security
      summary: Search profile-module summary
      operationId: searchProfileModules
      produces:
        - application/json
      parameters:
        - name: startDate
          in: query
          required: true
          type: string
          format: date
        - name: endDate
          in: query
          required: true
          type: string
          format: date
        - name: entityId
          in: query
          required: false
          type: string
        - name: entityName
          in: query
          required: false
          type: string
        - name: status
          in: query
          required: false
          type: integer
          format: int32
        - name: langId
          in: query
          required: false
          type: integer
          format: int32
      responses:
        '200':
          description: Profile-module summary list
          schema:
            $ref: '#/definitions/SecuritySummaryListResponse'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ExceptionWrapper'
      security:
        - bearerAuth: []
  /security/profile-module-processes:
    get:
      tags:
        - security
      summary: Search profile-module-process summary
      operationId: searchProfileModuleProcesses
      produces:
        - application/json
      parameters:
        - name: startDate
          in: query
          required: true
          type: string
          format: date
        - name: endDate
          in: query
          required: true
          type: string
          format: date
        - name: entityId
          in: query
          required: false
          type: string
        - name: entityName
          in: query
          required: false
          type: string
        - name: status
          in: query
          required: false
          type: integer
          format: int32
        - name: langId
          in: query
          required: false
          type: integer
          format: int32
      responses:
        '200':
          description: Profile-module-process summary list
          schema:
            $ref: '#/definitions/SecuritySummaryListResponse'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ExceptionWrapper'
      security:
        - bearerAuth: []
  /security/application-events:
    get:
      tags:
        - security
      summary: Search application-event catalog summary
      operationId: searchApplicationEvents
      produces:
        - application/json
      parameters:
        - name: startDate
          in: query
          required: true
          type: string
          format: date
        - name: endDate
          in: query
          required: true
          type: string
          format: date
        - name: entityId
          in: query
          required: false
          type: string
        - name: entityName
          in: query
          required: false
          type: string
        - name: status
          in: query
          required: false
          type: integer
          format: int32
        - name: langId
          in: query
          required: false
          type: integer
          format: int32
      responses:
        '200':
          description: Application-event summary list
          schema:
            $ref: '#/definitions/SecuritySummaryListResponse'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ExceptionWrapper'
      security:
        - bearerAuth: []
  /security/user-utility:
    post:
      tags:
        - security
      summary: Register or touch catalog user from JWT (utility entry)
      description: |
        Uses Authorization Bearer JWT; upserts core_security.user_data by claim `sub` or updates `updated_at` if the row exists. Intended once per client session when opening the utility home screen.
      operationId: postUserUtilitySession
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            type: object
            description: Optional empty JSON object; claims are taken from the Bearer token.
      responses:
        '200':
          description: User registered or catalog row updated
          schema:
            $ref: '#/definitions/GenericSuccessResponse'
        '400':
          description: Bad request (e.g. token without sub)
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '401':
          description: Unauthorized or invalid JWT
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ExceptionWrapper'
      security:
        - bearerAuth: []
  '/security/profiles/{id}/users':
    get:
      tags:
        - security
      summary: Get profile-user assignment
      operationId: getProfileUserAssignment
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          description: Profile identifier
          type: integer
          format: int32
      responses:
        '200':
          description: Available and assigned users for profile
          schema:
            $ref: '#/definitions/AssignmentResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '404':
          description: Profile not found
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ExceptionWrapper'
      security:
        - bearerAuth: []
    put:
      tags:
        - security
      summary: Save profile-user assignment
      operationId: saveProfileUserAssignment
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          description: Profile identifier
          type: integer
          format: int32
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/AssignmentSavePayload'
      responses:
        '200':
          description: Assignment updated
          schema:
            $ref: '#/definitions/GenericSuccessResponse'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '404':
          description: Profile not found
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ExceptionWrapper'
      security:
        - bearerAuth: []
  '/security/profiles/{id}/modules':
    get:
      tags:
        - security
      summary: Get profile-module assignment
      operationId: getProfileModuleAssignment
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          description: Profile identifier
          type: integer
          format: int32
        - name: langId
          in: query
          required: false
          type: integer
          format: int32
      responses:
        '200':
          description: Available and assigned modules for profile
          schema:
            $ref: '#/definitions/AssignmentResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '404':
          description: Profile not found
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ExceptionWrapper'
      security:
        - bearerAuth: []
    put:
      tags:
        - security
      summary: Save profile-module assignment
      operationId: saveProfileModuleAssignment
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          description: Profile identifier
          type: integer
          format: int32
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/AssignmentSavePayload'
      responses:
        '200':
          description: Assignment updated
          schema:
            $ref: '#/definitions/GenericSuccessResponse'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '404':
          description: Profile not found
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ExceptionWrapper'
      security:
        - bearerAuth: []
...
  /security/role-users:
    get:
      tags:
        - security
      summary: Search role-user summary
      operationId: searchRoleUsers
      produces:
        - application/json
      parameters:
        - name: startDate
          in: query
          required: true
          type: string
          format: date
        - name: endDate
          in: query
          required: true
          type: string
          format: date
        - name: entityId
          in: query
          required: false
          type: string
        - name: entityName
          in: query
          required: false
          type: string
        - name: status
          in: query
          required: false
          type: integer
          format: int32
        - name: langId
          in: query
          required: false
          type: integer
          format: int32
      responses:
        '200':
          description: Role-user summary list
          schema:
            $ref: '#/definitions/SecuritySummaryListResponse'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ExceptionWrapper'
      security:
        - bearerAuth: []
...
definitions:
  ExceptionWrapper:
    type: object
    properties:
      code:
        description: Numeric HTTP Status error
        example: 401
        format: int32
        type: integer
      message:
        description: Error description
        example: Unauthorized
        type: string
...
  SecuritySummaryListData:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: '#/definitions/SecuritySummaryRow'
      warningCode:
        type: string
      warningMessage:
        type: string
  SecuritySummaryListResponse:
    type: object
    properties:
      success:
        type: boolean
        example: true
      data:
        $ref: '#/definitions/SecuritySummaryListData'
...
  AssignmentResponse:
    type: object
    properties:
      success:
        type: boolean
        example: true
      data:
        $ref: '#/definitions/AssignmentData'
...
  GenericSuccessResponse:
    type: object
    properties:
      success:
        type: boolean
        example: true
      message:
        type: string
...
securityDefinitions:
  api_key:
    in: query
    name: key
    type: apiKey
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
  keycloak:
    authorizationUrl: ''
    flow: implicit
    scopes: {}
    type: oauth2
    x-google-audiences: vendor-backend
    x-google-issuer: '${KEYCLOAK}/auth/realms/corp'
    x-google-jwks_uri: '${JWKS_URL}'
    x-google-jwt-locations:
      - header: Authorization
        value_prefix: 'Bearer '
security:
  - bearerAuth: []
...